jekyll: ignore local gems
After "bundle config set --local path gems", bundler will install to the
local path gems, which is nice when you don't want to contaminate the
system (or run bundler with sudo/doas, which also requires tweaking the
Makefile).

However, this path is automatically included in the build, so exclude
it. I could probably choose to use vendor/gems instead, but let's try
this for now.

I also ended up doing

    print -l .bundle gems >>.git/info/exclude

for now, but they may want to go in .gitignore.